RSPBT v0.0.1
Work in progress Bot build inspired by the lore of Destiny 1&2 .

Feel free to rename the bot whatever you want, this is just a passion project I'm not really that pressed about it. It's not going to be a Dyno killer or anything. 

Contained in the .zip is
#########
rasputinbot.py
requirements.txt (ignore this, I've automated the python file to read this and install the appropriate version of python for this)
readme.bat (interactive Q&A batch)